+++ mmaction/datasets/__init__.py
@@ -1,4 +1,6 @@
 from .activitynet_dataset import ActivityNetDataset
+from .audio_dataset import AudioDataset
+from .audio_feature_dataset import AudioFeatureDataset
 from .base import BaseDataset
 from .builder import build_dataloader, build_dataset
 from .dataset_wrappers import RepeatDataset
@@ -8,5 +10,6 @@ from .video_dataset import VideoDataset
 
 __all__ = [
     'VideoDataset', 'build_dataloader', 'build_dataset', 'RepeatDataset',
-    'RawframeDataset', 'BaseDataset', 'ActivityNetDataset', 'SSNDataset'
+    'RawframeDataset', 'BaseDataset', 'ActivityNetDataset', 'SSNDataset',
+    'AudioDataset', 'AudioFeatureDataset'
 ]