+++ mmaction/datasets/pipelines/__init__.py
@@ -1,15 +1,16 @@
-from .augmentations import (CenterCrop, ColorJitter, Flip, Fuse,
-                            MultiGroupCrop, MultiScaleCrop, Normalize,
-                            RandomCrop, RandomResizedCrop, Resize, TenCrop,
-                            ThreeCrop)
+from .augmentations import (AudioAmplify, CenterCrop, ColorJitter, Flip, Fuse,
+                            MelSpectrogram, MultiGroupCrop, MultiScaleCrop,
+                            Normalize, RandomCrop, RandomResizedCrop, Resize,
+                            TenCrop, ThreeCrop)
 from .compose import Compose
-from .formating import (Collect, FormatShape, ImageToTensor, ToDataContainer,
-                        ToTensor, Transpose)
-from .loading import (DecordDecode, DecordInit, DenseSampleFrames,
+from .formating import (Collect, FormatAudioShape, FormatShape, ImageToTensor,
+                        ToDataContainer, ToTensor, Transpose)
+from .loading import (AudioDecode, AudioDecodeInit, AudioFeatureSelector,
+                      DecordDecode, DecordInit, DenseSampleFrames,
                       FrameSelector, GenerateLocalizationLabels,
-                      LoadLocalizationFeature, LoadProposals, OpenCVDecode,
-                      OpenCVInit, PyAVDecode, PyAVInit, RawFrameDecode,
-                      SampleFrames, SampleProposalFrames,
+                      LoadAudioFeature, LoadLocalizationFeature, LoadProposals,
+                      OpenCVDecode, OpenCVInit, PyAVDecode, PyAVInit,
+                      RawFrameDecode, SampleFrames, SampleProposalFrames,
                       UntrimmedSampleFrames)
 
 __all__ = [
@@ -20,6 +21,7 @@ __all__ = [
     'Collect', 'FormatShape', 'Compose', 'ToTensor', 'ToDataContainer',
     'GenerateLocalizationLabels', 'LoadLocalizationFeature', 'LoadProposals',
     'DecordInit', 'OpenCVInit', 'PyAVInit', 'SampleProposalFrames',
-    'UntrimmedSampleFrames', 'RawFrameDecode', 'DecordInit', 'OpenCVInit',
-    'PyAVInit', 'SampleProposalFrames', 'ColorJitter'
+    'UntrimmedSampleFrames', 'RawFrameDecode', 'AudioAmplify',
+    'MelSpectrogram', 'AudioDecode', 'FormatAudioShape', 'ColorJitter',
+    'LoadAudioFeature', 'AudioFeatureSelector', 'AudioDecodeInit'
 ]