+++ mmaction/models/__init__.py
@@ -4,12 +4,14 @@ from .backbones import (ResNet, ResNet2Plus1d, ResNet3d, ResNet3dCSN,
 from .builder import (build_backbone, build_head, build_localizer, build_loss,
                       build_model, build_neck, build_recognizer)
 from .common import Conv2plus1d
-from .heads import BaseHead, I3DHead, SlowFastHead, TPNHead, TSMHead, TSNHead
+from .heads import (AudioTSNHead, BaseHead, I3DHead, SlowFastHead, TPNHead,
+                    TSMHead, TSNHead)
 from .localizers import BMN, PEM, TEM
 from .losses import (BCELossWithLogits, BinaryLogisticRegressionLoss, BMNLoss,
                      CrossEntropyLoss, NLLLoss, OHEMHingeLoss, SSNLoss)
 from .necks import TPN
-from .recognizers import BaseRecognizer, recognizer2d, recognizer3d
+from .recognizers import (AudioRecognizer, BaseRecognizer, recognizer2d,
+                          recognizer3d)
 from .registry import BACKBONES, HEADS, LOCALIZERS, LOSSES, RECOGNIZERS
 
 __all__ = [
@@ -21,5 +23,5 @@ __all__ = [
     'BCELossWithLogits', 'LOCALIZERS', 'build_localizer', 'PEM', 'TEM',
     'BinaryLogisticRegressionLoss', 'BMN', 'BMNLoss', 'build_model',
     'OHEMHingeLoss', 'SSNLoss', 'ResNet3dCSN', 'ResNetTIN', 'TPN', 'TPNHead',
-    'build_loss', 'build_neck'
+    'build_loss', 'build_neck', 'AudioRecognizer', 'AudioTSNHead'
 ]